requirements.txt
numpy>=2.1 
pandas>=2.2.3   
duckdb>=1.1        
pyyaml==6.0.2
python-dotenv==1.0.1
streamlit>=1.37.0
regex==2024.5.15